+++ src/utilities/CartesianMesh.h
@@ -8,6 +8,8 @@
 #if !defined(CARTESIAN_MESH_H)
 #define CARTESIAN_MESH_H
 
+#include "types.h"
+
 #include <string>
 #include <vector>
 
@@ -39,6 +41,11 @@ public:
   ~CartesianMesh();
   // parse input file and create Cartesian mesh
   void initialize(std::string filePath);
+  // write grid points into file
+  PetscErrorCode write(std::string filePath);
+#ifdef PETSC_HAVE_HDF5
+  PetscErrorCode write(std::string filePath, StaggeredMode mode);
+#endif
   // print information about Cartesian mesh
   PetscErrorCode printInfo();