+++ src/utilities/SimulationParameters.cpp
@@ -58,6 +58,20 @@ void SimulationParameters::initialize(std::string filePath)
   startStep = node["startStep"].as<PetscInt>(0);
   nt = node["nt"].as<PetscInt>();
   nsave = node["nsave"].as<PetscInt>(nt);
+  
+  outputFormat = node["outputFormat"].as<std::string>("binary");
+#ifndef PETSC_HAVE_HDF5
+  if (outputFormat == "hdf5")
+  {
+    PetscPrintf(PETSC_COMM_WORLD,
+                "\nERROR: PETSc has not been built with HDF5 available; "
+                "you cannot use `outputFormat: hdf5`\n");
+    MPI_Barrier(PETSC_COMM_WORLD);
+    exit(1);
+  }
+#endif
+  outputFlux = (node["outputFlux"].as<bool>(true)) ? PETSC_TRUE : PETSC_FALSE;
+  outputVelocity = (node["outputVelocity"].as<bool>(false)) ? PETSC_TRUE : PETSC_FALSE;
 
   ibm = stringToIBMethod(node["ibm"].as<std::string>("NONE"));
 
@@ -141,6 +155,9 @@ PetscErrorCode SimulationParameters::printInfo()
   ierr = PetscPrintf(PETSC_COMM_WORLD, "starting time-step: %d\n", startStep); CHKERRQ(ierr);
   ierr = PetscPrintf(PETSC_COMM_WORLD, "number of time-steps: %d\n", nt); CHKERRQ(ierr);
   ierr = PetscPrintf(PETSC_COMM_WORLD, "saving-interval: %d\n", nsave); CHKERRQ(ierr);
+  ierr = PetscPrintf(PETSC_COMM_WORLD, "output format: %s\n", outputFormat.c_str()); CHKERRQ(ierr);
+  ierr = PetscPrintf(PETSC_COMM_WORLD, "output flux: %D\n", outputFlux); CHKERRQ(ierr);
+  ierr = PetscPrintf(PETSC_COMM_WORLD, "output velocity: %D\n", outputVelocity); CHKERRQ(ierr);
   ierr = PetscPrintf(PETSC_COMM_WORLD, "---------------------------------------\n"); CHKERRQ(ierr);
 
   return 0;