+++ src/utilities/SimulationParameters.h
@@ -43,6 +43,10 @@ public:
            nt,        ///< number of time steps
            nsave;     ///< data-saving interval
   
+  std::string outputFormat;  ///< output format to use
+  PetscBool outputFlux,     ///< boolean to output the flux components
+            outputVelocity; ///< boolean to output the velocity components
+
   IBMethod ibm; ///< type of system to be solved
   
   TimeIntegration convection, ///< time-scheme for the convection term